requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

python=3.10.13
torch=2.0.1
torch-summary=1.4.5
torchvision=0.15.2
numpy=1.24.4
opencv-contrib-python=4.8.1.78
opencv-python=4.10.0.84
ultralytics=8.3.57
tqdm=4.65.2
yaml=0.2.5
matplotlib=3.5.3
seaborn=0.12.2
qt-main=5.15.2
qtconsole=5.5.0
qtpy=2.4.1
pyqt=5.15.10=
pyqt5-sip=12.13.0
pillow=0.1.0
